Fix a crash when app exits after Map is manually destroyed with a non-empty JobQueue
Once all Map objects are destructed, any Jobs in JobQueue will remain there. The disposal jobs from
TileWorker holds a unique_ptr to a TileBuilder which has either itself been deleted or holds references
to deleted objects. When the app exits, globals are destructed and the destructor for the TileWorker
disposal jobs try to destruct that unique_ptr, resulting in a segfault.

This change removes the unique_ptr binding, relying instead on an explicit delete to the pointer. This
means that the delete will only happen when the job is run, not when it's destroyed. There is no potential
memory leak here because this only occurs when the process and all associated memory is being destroyed.